samples: use build-helper-maven-plugin to share sample/test code for different install methods
